feat(create pki): add command create pki to generate PKI files needed for on-prem clusters (#525)
This is a port of a piece of functionality currently present on
furyagent. The code has been modified and simplified to cover only the
PKI creation case, original code on furyagent supported much more
options like backups and automatically joining nodes in AWS.

Add a create pki command with options to create the Public Key
Infrastructure needed for on-premises Kubernetes clusters (etcd CA and
certs, Kubernetes control plane CA and certs).

The command allows limiting which PKI to create (etcd only, control
plane only or both) and also in which path to save the files.